Extract processRoadNetwork function. DRY
